src/requirements.txt
dill==0.3.6
gdown==4.7.1
GitPython==3.1.31
lightgbm==3.3.5
numpy==1.24.2
pycm==3.8
requests==2.28.2
ruamel.base==1.0.0
scikit_learn==1.2.2
torch==1.12.0+cu116
torchvision==0.13.0+cu116
